+++ lua/plugins/config/easy_align.lua
@@ -11,5 +11,6 @@ return function(use)
 				gA = { '<Plug>(EasyAlign)', '(TextObject) EasyAlign' },
 			}, { mode = 'n' })
 		end,
+		event = 'BufEnter',
 	})
 end