+++ lua/plugins/config/nvim_tree.lua
@@ -103,5 +103,6 @@ return function(use)
 		setup = setup,
 		config = run,
 		wants = { 'which-key.nvim' },
+		event = 'VimEnter',
 	})
 end